requirements.txt
pymongo==4.6.1
dnspython==2.4.2
python-dotenv==1.0.0
schedule==1.2.0
psutil==5.9.6
netifaces==0.11.0
requests==2.31.0
